requirements.txt
Flask==3.0.0
mysql-connector-python==9.0.0
flask-cors==4.0.0
gunicorn==23.0.0
